Start with old and broad product expectations
This is a slightly altered version of the expectations
collected from PMs a week or so ago.

As we move forward we can see if these continue to make sense or
if they should be moved to other stakeholders etc.

Ultimately these expectations have driven most of the wikidata
and wikibase development.
There are various usecases that have been created within these
broad topics, for example the "The stored data and changes must
be managable by a distributed community." hides many features
within it.