Split stakeholders and expectations into sub systems
Push many stakeholders and expectations into either
repo or client, some remain at the top level.

This allowed me to tweak some of the stakeholders and
expectations, also generalizing some of them away from
wikipedia and wikidata, toward repo and client where appropriate?
This probably needs confirming and review from product in the
future.

I also switched from - to * for unordered lists as my markdown lint
was raging and I think we should enable linting at some point ( so
as to provide the most consistent documentation as possible).

Possibly ALL of the lower level stakeholders are also high level
stakeholders and should also be listed? Or linked to?
We should think about this.