Use full extension names (not repo and client)
From CR:
It will be better to name the components WikibaseClient and WikibaseRepo
given that one of the first tripping points is the difference between
them and what is "client" in our context.
Also something like WikibaseLexeme will be added under that name
and not just "lexeme".
The component names will be technical and not conceptual.