Split the docs into multiple subsystems (#8)
* Add a blank copy of the arc42 template

As we slowly document more components, this can be used as a source
of truth.
Having the template here means we can also slowly tweak things as we
progress and learn in the context of arc42 that will not be lost for
future components.

* Add an empty template for "client"

We basically decided that we want to document this
component, so use it as an example for the proposed structure.

* Linked contents in template README file

* Introduce viewers to the component split in README

* Also add blank template for "repo"

I guess as we decided client will be its own thing repo
will also end up needing its own doc template.

This will also allow me to "play around with" moving out some
stakeholders and expectations into these components.

* Split our 1 quality goal into repo and client

* Split stakeholders and expectations into sub systems

Push many stakeholders and expectations into either
repo or client, some remain at the top level.

This allowed me to tweak some of the stakeholders and
expectations, also generalizing some of them away from
wikipedia and wikidata, toward repo and client where appropriate?
This probably needs confirming and review from product in the
future.

I also switched from - to * for unordered lists as my markdown lint
was raging and I think we should enable linting at some point ( so
as to provide the most consistent documentation as possible).

Possibly ALL of the lower level stakeholders are also high level
stakeholders and should also be listed? Or linked to?
We should think about this.

* Refer to ## Stakeholders & Expectations in all templates

* Switch to - instead of * for lists

I thought my linter was telling me *, but i was wrong.

* Fix wording in main README

* Use full extension names (not repo and client)

From CR:
It will be better to name the components WikibaseClient and WikibaseRepo
given that one of the first tripping points is the difference between
them and what is "client" in our context.
Also something like WikibaseLexeme will be added under that name
and not just "lexeme".
The component names will be technical and not conceptual.

* main: link to repo and client from 03-context

* Wikidata Researcher -> Researcher

"sounds like a person doing research about Wikidata to me?"

* components -> subsystems to disambiguate

* Fix subsystem paths

* Reflect file name spaces in README links

Maybe the file names should be changed to underscore instead, but I'm
leaving them intact for now to avoid conflicts.

Co-authored-by: sihe <silvan.heintze@wikimedia.de>