Copy top level list of systems and sub systems
This comes from a Google doc (with slighlt name changes).

Even if we tweak this in the future I think we would benefit
having this list in Git now.

It could be that in the future we decide to get rid of this home
page and instead only have the overview?