Turn Usage Tracking into a top level building block
How Entities are used on Client pages is tracked by storing usage
information in the client's database whenever an edit to Wikitext
referencing Entity data is made. This should be noted as a communication
channel in the technical context diagram, and should probably also make
Usage Tracking a top level building block.