Start a whitebox overall system diagram
Motivated by the fact that I want to add some diagrams related
to the current state of the quality constraints component of repo.
This Ideally I need a starting point top level diagram that I can
ass this block to.

This diagram currently focuses on the interactions between external
systems and users and the repo.
There is a note saying this diagram is incomplete!